cleaned up the tests

+++ app/helpers/application_helper.rb
@@ -1,7 +1,7 @@
 module ApplicationHelper
 
     def self.pct value, total
-        value / total * 100
+        Integer(Float(value) / Float(total) * 100)
     end
 
 end
